requirements.txt
flask
flask-bcrypt
pymongo
python-dotenv
openai
spacy
python-docx
PyPDF2
python-pptx
Pillow
pytesseract
rapidfuzz